Median - 2nd version - for the whole

+++ scala/src/main/scala/SaturationWithHistogramForAll.scala
@@ -201,7 +201,7 @@ object SaturationWithHistogramForAll {
           .over(Window.partitionBy("group").orderBy($"label")))
         .withColumn("start", (col("end") - col("count")))
 
-      var zeros = histogram.filter($"label" === 0).select("count").first().getInt(0)
+      var zeros = histogram.filter($"label" === 0).select("count").first().getLong(0)
       zerosRow = zerosRow :+ (zeros * 100 / total)
 
       if (isImpair) {
